+++ mod/twitch/src/main/java/tv/twitch/android/models/streams/StreamModel.java
@@ -1,0 +1,9 @@
+package tv.twitch.android.models.streams;
+
+public class StreamModel {
+  public String createdAt;
+
+  public tv.twitch.android.models.streams.StreamType getStreamType() {
+    throw new IllegalStateException("tv.twitch.android.models.streams.getStreamType() stub was called");
+  }
+}