+++ mod/twitch/src/main/java/tv/twitch/android/shared/player/overlay/BottomPlayerControlOverlayViewDelegate.java
@@ -1,0 +1,8 @@
+package tv.twitch.android.shared.player.overlay;
+
+import tv.twitch.android.models.streams.StreamType;
+
+public class BottomPlayerControlOverlayViewDelegate {
+  public android.widget.TextView mLiveIndicatorLeftText;
+  public void updateStreamType(StreamType streamType) {}
+}